pyproject.toml
[project]
name = "brest-mcp"
version = "0.1.0"
description = "A MCP server project for the city of Brest that enables access to weather, transports & events data feeds"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.4.1",
    "mcp[cli]",
    "pydantic>=2.0.0",
    "gtfs-realtime-bindings>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "folium>=0.15.0",
    "plotly>=5.18.0",
    "plotly-express>=0.4.1",
    "streamlit-folium>=0.18.0",
    "streamlit>=1.32.0",
    "ollama>=0.1.6",
    "anthropic>=0.50.0",
    "langchain-mcp-adapters>=0.1.1",
    "langgraph-prebuilt>=0.1.8",
    "langgraph>=0.4.5",
    "langchain-openai>=0.3.17",
    "a2a-sdk>=0.2.4",
    "langchain-anthropic>=0.3.12",
]
[[project.authors]]
name = "Artemis-IA"
email = "132653903+Artemis-IA@users.noreply.github.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
brest-mcp = "src:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
